+++ setUbuntuProxyPacFile.bash
@@ -1,3 +1,0 @@
-#!/usr/bin/env bash
-
-gsettings set org.gnome.system.proxy autoconfig-url '$1'