+++ src/main/java/org/lantern/StatsTrackingHandler.java
@@ -6,7 +6,6 @@ import org.jboss.netty.channel.ChannelHandlerContext;
 import org.jboss.netty.channel.MessageEvent;
 import org.jboss.netty.channel.SimpleChannelHandler;
 import org.jboss.netty.channel.WriteCompletionEvent;
-
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;