+++ src/main/java/org/lantern/linux/Unique.java
@@ -1,8 +1,6 @@
 package org.lantern.linux;
 
-import com.sun.jna.Callback;
 import com.sun.jna.Library;
-import com.sun.jna.Native;
 import com.sun.jna.Pointer;
 
 public interface Unique extends Library {