+++ dashboard/assets/CHANGES.md
@@ -1,11 +1,42 @@
 Lantern Changelog
 =================
 
+
+Version 0.97.7
+--------------
+
+Released 2012-07-06. [Compare
+view](https://github.com/getlantern/lantern/compare/v0.97.6...v0.97.7)
+
+- Implemented set as system proxy and start on login for Ubuntu 12.04
+
+
+Version 0.97.6
+--------------
+
+Released 2012-04-26. [Compare
+view](https://github.com/getlantern/lantern/compare/v0.97.5...v0.97.6)
+
+
+Version 0.97.5
+--------------
+
+Released 2012-04-13. [Compare
+view](https://github.com/getlantern/lantern/compare/v0.97.4...v0.97.5)
+
+
+Version 0.97.4
+--------------
+
+Released 2012-03-29. [Compare
+view](https://github.com/getlantern/lantern/compare/v0.97.3...v0.97.4)
+
+
 Version 0.97.3
 --------------
 
-Release date to be determined.
-
+Released 2012-. [Compare
+view](https://github.com/getlantern/lantern/compare/v0.97.2...v0.97.3)
 
 - Restored Windows and Linux support
 
@@ -15,26 +46,19 @@ Release date to be determined.
   dashboard
 
 
-GitHub compare view: [http://git.io/Bfj1Lw](https://github.com/getlantern/lantern/compare/v0.97.2...v0.97.3)
-
-
 Version 0.97.2
 --------------
 
-Released 2012-02-17.
+Released 2012-02-17. [Compare
+view](https://github.com/getlantern/lantern/compare/v0.97.1...v0.97.2)
 
-
-- Fixed an issue running Lantern on Mac OS X 10.5 (#170).
-
-
-GitHub compare view: [http://git.io/O_RrUQ](https://github.com/getlantern/lantern/compare/v0.97.1...v0.97.2)
+- Fixed an issue running Lantern on Mac OS X 10.5 (#170)
 
 
 Version 0.97.1
 --------------
 
 Released 2012-02-12.
-
 
 - UI now implemented in the new Lantern Dashboard. Allows for displaying
   activity, modifying configuration, and providing feedback.